refactor(api.analyze): restructure analyze module with helpers, protocols, and result classes
- Extract file and scene update check logic to analyze_helpers.py
- Define SceneAnalyzer protocol in analyze_protocols.py
- Move FileResult and AnalyzeResult dataclasses to analyze_results.py
- Update AnalyzeCommand to use new helpers and protocols
- Simplify imports and remove redundant code
- Adjust tests to import from new modules

This refactor improves code organization, readability, and maintainability by separating concerns and defining clear interfaces for analyzers.
