+++ src/scriptrag/api/__init__.py
@@ -1,6 +1,7 @@
 """ScriptRAG API module."""
 
-from scriptrag.api.analyze import AnalyzeCommand, AnalyzeResult, FileResult
+from scriptrag.api.analyze import AnalyzeCommand
+from scriptrag.api.analyze_results import AnalyzeResult, FileResult
 from scriptrag.api.database import DatabaseInitializer
 from scriptrag.api.database_operations import DatabaseOperations, ScriptRecord
 from scriptrag.api.index import IndexCommand, IndexOperationResult, IndexResult
